+++ src/shared/errors/index.js
@@ -1,15 +1,6 @@
 import { Alert } from 'react-native';
 
-// eslint-disable-next-line no-underscore-dangle
-const isDev = global.__DEV__;
-
-const defaultHandler = global.ErrorUtils.getGlobalHandler();
-global.ErrorUtils.setGlobalHandler((error, isFatal) => {
-  if (isDev) {
-    defaultHandler(error, isFatal);
-    return;
-  }
-
+const productionErrorHandler = (error, isFatal) => {
   /* eslint-disable no-console */
   console.log(`Caught JS error: (${isFatal ? 'fatal' : 'non-fatal'})`);
   console.log(error);
@@ -20,4 +11,9 @@ global.ErrorUtils.setGlobalHandler((error, isFatal) => {
     error.message,
     [{ text: 'OK' }],
   );
-});
+};
+
+// eslint-disable-next-line no-underscore-dangle
+if (!global.__DEV__) {
+  global.ErrorUtils.setGlobalHandler(productionErrorHandler);
+}