+++ docs/CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+* When discarding a work session, show a Toast with feedback
 
+### Fixed
+* Item overflow in subjects list
 
 
 ## 1.9.1 - 2019-10-29